moves code around to avoid issues with circular dependencies and documents changes to the router
