+++ src/common/app-link/index.ts
@@ -1,0 +1,1 @@
+export * from './app-link';
