+++ src/common/app-redirect/index.ts
@@ -1,0 +1,1 @@
+export * from './app-redirect';
