+++ src/pages/home/home.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 
 import globalStyles from 'assets/stylesheets/global-styles.module.scss';
-import { AppLink } from 'routes/app-link';
+import { AppLink } from 'common/app-link';
 import { RouteName } from 'routes/routes';
 import { Button } from 'common/button';
 import { Avatar } from 'common/avatar';
