+++ src/pages/not-found/not-found.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { AppLink } from 'routes/app-link';
+import { AppLink } from 'common/app-link';
 import { RouteName } from 'routes/routes';
 
 // TODO: implement this page
