+++ src/routes/index.ts
@@ -1,5 +1,4 @@
-export { AppLink } from './app-link';
-export { AppRedirect } from './app-redirect';
-export { goToPage } from './route-helpers';
+export { getRouteFor } from './route-helpers';
 export { routes, RouteName } from './routes';
+export type { Params } from './routes';
 export { createRouter } from './router';
