+++ src/routes/route-helpers.ts
@@ -1,4 +1,4 @@
-import { redirect } from 'react-router-dom';
+import { generatePath } from 'react-router-dom';
 import { ParamsHelper } from 'helpers/params-helper';
 import { routes, Params, RouteName } from './routes';
 
@@ -23,33 +23,14 @@ const getRouteFor = (routeName: RouteName, pathParams: Params = {}, queryParams:
   const route = findRoute(routeName);
 
   // Replace pathParams
-  const routePathParams = route.path.split('/');
-  const routePathPms = routePathParams.map((pathParam) => {
-    if (pathParam.startsWith(':')) {
-      const paramName = pathParam.match(/:[^/?]+/gi)?.[0].slice(1);
-      if (!paramName) {
-        throw new Error(`Path parameter '${pathParam}' is not valid for route '${routeName}'`);
-      }
-      return pathParams[paramName] || '';
-    }
-    return pathParam;
-  });
-
-  const routePath = `/${routePathPms.filter((pathParam) => pathParam !== '').join('/')}`;
+  const interpolatedPath = generatePath(route.path, pathParams);
 
   // Inject queryParams
   const urlQueryParams = ParamsHelper.createQueryParams(queryParams);
-  return `${routePath}${urlQueryParams}`;
-};
-
-const goToPage = (routeName: RouteName, pathParams: Params = {}, queryParams: Params = {}) => {
-  const routePath = getRouteFor(routeName, pathParams, queryParams);
-
-  redirect(routePath);
+  return `${interpolatedPath}${urlQueryParams}`;
 };
 
 export {
   findRoute,
   getRouteFor,
-  goToPage,
 };
